Update Trajectories Modal: - Refactor the trajectories modal to be more UX friendly by introducing better icons, labels, accordions, etc. - Extract value map as a separate diagram component to be used in the value map page and trajectory. Is collapsible. - Include trajectory update migration and update client function to instruct agent on how to process query.
