+++ backend/app/alembic/versions/a20e49758210_update_trajectory_questions.py
@@ -1,0 +1,33 @@
+"""update_trajectory_questions
+
+Revision ID: a20e49758210
+Revises: 3fb943dcb6f5
+Create Date: 2026-04-23 21:15:35.321915
+
+"""
+from alembic import op
+import sqlalchemy as sa
+import sqlmodel.sql.sqltypes
+
+
+# revision identifiers, used by Alembic.
+revision = 'a20e49758210'
+down_revision = '3fb943dcb6f5'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('trajectory', sa.Column('rephrased_morning_question', sa.Text(), nullable=True))
+    op.add_column('trajectory', sa.Column('rephrased_evening_question', sa.Text(), nullable=True))
+    op.drop_column('trajectory', 'rephrased_question')
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('trajectory', sa.Column('rephrased_question', sa.TEXT(), autoincrement=False, nullable=True))
+    op.drop_column('trajectory', 'rephrased_evening_question')
+    op.drop_column('trajectory', 'rephrased_morning_question')
+    # ### end Alembic commands ###
